Default recurrent_compute to fp16: H256 9.1->7.4s (-19%), H512 41->32s (-22%)

+++ miopen_adaptive_lstm_hip/miopen_adaptive_lstm_hip/model.py
@@ -175,7 +175,7 @@ def _use_direct_blas() -> bool:
 
 
 def _recurrent_compute_mode() -> str:
-    raw = os.environ.get("MIOPEN_ADAPTIVE_LSTM_RECURRENT_COMPUTE", "fp32").strip().lower()
+    raw = os.environ.get("MIOPEN_ADAPTIVE_LSTM_RECURRENT_COMPUTE", "fp16").strip().lower()
     if raw in {"", "fp32", "32", "float"}:
         return "fp32"
     if raw in {"fp16", "16", "half"}:
